removes hsbutton in favor of clbsbutton
